Refactor: remove all code from mod.rs and lib.rs
Refactor the sdk to put all "real" code in named files,
with only module definitions like `pub mod my_mod;`
in the lib.rs and mod.rs files. This makes it easier for
newcomers to the codebase to navigate
